refactored so that base widget can propagate buffer changes
